requirements.txt
torch>=2.5.0
torchvision>=0.20.0
diffusers>=0.35.0
transformers>=4.47.0
accelerate>=1.2.1
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
gradio>=5.12.0
pillow>=10.4.0
opencv-python>=4.10.0.84
imageio>=2.36.1
imageio-ffmpeg>=0.5.1
numpy>=1.26.4
safetensors>=0.4.5
python-multipart>=0.0.20
xformers>=0.0.29.post3; platform_system == "Linux" and platform_machine == "x86_64"
